Extract installers as an interface
